Use the tarSingleFileToExtract feature for krew
I was hoping to not have to use this feature, but there are at least
three projects that use this odd convention to name the tarball with
the os and arch and then naming the binary within using the os and
arch. Krew, Velero, and Yq use this convention.

Specifying it in the tool descriptor allows us not to have special
code in post-install scripts to rename the file.