Structure Graal like Temurin to prepare having one Java/JDK config
Graal is just another JDK, but it's downloaded from a different location
and is packaged in a slightly different way. This is likely true of
many of the JDK options so we might need to create an option where
a pointer to a script to provide the download location as interpolating
the version, os, and architecture is likely not enough.